when embedding in Koha, do not require user's password
Part of work to make biblios embedded in Koha automatically
authenticate via the Koha user session cookie, this
patch adds a new Koha web service, proxy_auth_cookie,
which when passed the session cookie of a valid Koha
session, clones the session and returns the new
cookie as PROXYSESSID.  The trick is that since
biblios.pl is invoking proxy_auth_cookie via LWP::UserAgent,
the new session that is created has the correct IP address, i.e.,
that of the Koha server.

biblios.pl then passes the new cookie via embeddedSESSID; when
this cookie is given to the proxy script kohaws.pl, kohaws.pl
will no longer need to know the operator's password.

To complete this work, the following will need to be done:

1. kohaws.pl should be modified so that it knows when it
   is being passed a valid proxy cookie, signalling that it
   is being used by a biblios client embedded in a Koha
   installation.
2. biblios should be modified to not ask for an operator
   password for the Koha send target when it is being
   run in embedded mode.

Note that proxy_auth_cookie should be installed in the svc/
subdirectory of the Koha staff CGI directory.

One assumption being made, which should generally be accurate
in the case of embedded biblios but may require tweaking in certain
split-server setups, is that kohaws.pl is run from the same webserver
as biblios.pl and proxy_auth_cookie.

[LL bug 310]